added Dilemma Soul Devil (Spirit Eater Support)
3+ Fiend and or Zombie monsters, including a "Spirit Eater" monster
(This card is always treated as a "Spirit Eater" card.)
Monsters this card points to gain 1500 ATK. This card cannot be destroyed by monsters it points to (by battle or card effects). Twice per turn, if your opponent Summons a monster(s) to a zone this card points to: Draw 1 card for each of those monsters, then, if you drew more than 1 card, discard cards equal to the number of cards you drew minus 1. Once per turn: You can target 1 Special Summoned monster this card does not point to; destroy it. This is a Quick Effect if you control a Fiend or Zombie monster other than this card.