Added "Outer Entity Chrohel" and moved "Palor"-Folder into "script"-Folder
3 Level 3 or higher monster with the same Level
Once per turn: You can detach 1 material from this card; negate the effects of all other face-up monsters currently on the field until the end of the next turn. Unaffected by Spell/Trap effects that are activated in the same column as this card. Each time exactly 1 monster is Normal or Special Summoned to your opponent’s field: Move this card to the adjacent Main Monster Zone closer to that opponent’s monster, then attach all other monsters in this card’s column to this card.
[ Artwork by: Storm Wraith - https://www.deviantart.com/morkardfc ]
[ Expanded to 1:1 aspect ratio with AI - https://www.pixelcut.ai/uncrop ]